+++ NEWS.md
@@ -1,5 +1,6 @@
 # rsofun (development version)
 * BiomeEP:
+  * Bugfix: `phenotype` is now correctly taking into account, decoupling it from `LMA` (#329)
   * Bugfix: annual, cohort-level output had mixed up column names for variables:
   `NSN`,`seedC`,`leafC`,`rootC`,`sapwoodC`,`heartwoodC`
   * Bugfix: `init_cohort$init_cohort_nsc` is now correctly taken into account. To 
@@ -11,7 +12,7 @@
   * Cohorts are now less aggressively merged. Merging criteria was simplified from
   relative to absolute DBH difference. Now merging if difference <= 0.01 m.
   * Removed dummy parameters in `params_species` for `run_biomee_f_bysite()`: 
-  `phenotype`,`Vmax`,`alphaBM`,`leafLS`,`lAImax`,`CNleaf0`,`gamma_L`,`Vannual`,
+  `Vmax`,`alphaBM`,`leafLS`,`lAImax`,`CNleaf0`,`gamma_L`,`Vannual`,
   `betaON`,`betaOFF`, `leaf_size` and in `params_tile`: `GR_factor`.
   If still provided, they must be NA, otherwise an error occurs.
   * Removed parameter in `params_tile`  for `run_biomee_f_bysite()`: `par_mort_under`